refactor(agents): extract token usage aggregator
